用户登录完成

+++ src/modules/doctorIntroduction/doctorIntroductionView.js
@@ -1,8 +1,9 @@
 (function(app) {
   'use strict';
 
-  var doctorIntroductionViewCtrl = function($scope, $http, $state, $stateParams, $cordovaToast) {
+  var doctorIntroductionViewCtrl = function($scope, $http, $state, $stateParams, $cordovaToast, userService) {
     $scope.type = ($stateParams.type==='1');
+    var isLogin = userService.hasToken();
 
     //取得医生简介
     $http.get('/doctors/'+$stateParams.doctorId).success(function(data) {
@@ -24,30 +25,44 @@
       else {
         $scope.isCollection = true;
       }
-    }).error(function(data){
-      $cordovaToast.showShortBottom(data);
-    });
-    $scope.collectionDoctor = function() {
-      if ($scope.isCollection) {
-        $http.delete('/user/collectionDoctors/'+$stateParams.doctorId).success(function() {
-          $scope.isCollection = false;
-        }).error(function(data){
-          $cordovaToast.showShortBottom(data);
-        });
+    }).error(function(data, status){
+      if (status === 401) {
+        $scope.isCollection = false;
       }
       else {
-        $http.post('/user/collectionDoctors', {doctorId: $stateParams.doctorId}).success(function() {
-          $scope.isCollection = true;
-        }).error(function(data){
-          $cordovaToast.showShortBottom(data);
-        });
+        $cordovaToast.showShortBottom(data);
       }
-
+    });
+    $scope.collectionDoctor = function() {
+      if (isLogin) {
+        if ($scope.isCollection) {
+          $http.delete('/user/collectionDoctors/'+$stateParams.doctorId).success(function() {
+            $scope.isCollection = false;
+          }).error(function(data){
+            $cordovaToast.showShortBottom(data);
+          });
+        }
+        else {
+          $http.post('/user/collectionDoctors', {doctorId: $stateParams.doctorId}).success(function() {
+            $scope.isCollection = true;
+          }).error(function(data){
+            $cordovaToast.showShortBottom(data);
+          });
+        }
+      }
+      else {
+        $state.go('login');
+      }
     };
 
     //挂号点击事件
     $scope.register = function() {
-      $state.go('registerDoctorTimeSelect', {doctorId: $stateParams.doctorId, date: ''});
+      if (isLogin) {
+        $state.go('registerDoctorTimeSelect', {doctorId: $stateParams.doctorId, date: ''});
+      }
+      else {
+        $state.go('login');
+      }
     };
   };
